+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+## WIP (0.1.1)
+
+- Showing window when URL have finished Loading
+- Added an error when the URL fails to load for any reason
+
 ## 0.1.0
 
 - Initial proof of concept